Module Ppx_expect_runtime.Compact_loc

type t = {

start_bol : Base.int;
(*
Index of the first character of the first line in the range.
*)

start_pos : Base.int;
(*
Index of the first character in the range.
*)

end_pos : Base.int;
(*
Index of the last character in the range.
*)

}

A range of characters in a source file.

Notably, Compact_loc.t does not store the name of the file itself; it only represents the range of characters.

Consider this diagram, where the characters /[a-b]/ represent the included range, and the characters /[.0]/ represent the surrounding text:

.......
0.........a-----
------------
-----b
.........

Then start_bol is the position of 0, start_pos of a, and end_pos of b.

This record corresponds to a Ppxlib.Location.t restricted to just the loc_start.pos_cbol, loc_start.pos_cnum, and loc_end.pos_cnum fields.

val equal : t -> t -> Base.bool

Compact_loc.t satisfies the natural definition of equality. We write it by hand only to minimize external dependencies.

val compare_character_range : t -> t -> Base.int

Compares the range of characters spanned by a Compact_loc.t. Ranges that start earlier are considered smaller. For ranges that start at the same position, ranges that end earlier are considered smaller.