Module Ppx_expect_runtime.Expect_node_formatting

type t = {

indent : Base.int;
(*
The number of spaces that the bodies of [%expect] nodes are indented with respect to the left edge of the extension point and that trailing [%expect] nodes are indented with respect to the enclosing let%expect
*)

always_on_own_line : Base.bool;
(*
Whether the output of [%expect] nodes should always be formatted so it gets its own lines, even if it is only one line long. If this option is true, one-line expectations will be printed like

[%expect {|
  foo
|}]

rather than like

[%expect {| foo |}]
*)

extension_sigil : Base.string;
(*
The sigil that should be printed to signal the start of an extension point. By default, this is "%", though in toplevel tests it changes to "%%".
*)

attribute_sigil : Base.string;
(*
The sigil that should be printed to signal the start of an attribute. By default, this is "@@".
*)

}

Configurations for the formatting of rewritten expect nodes and attributes. The values in default are used by ppx_expect, but different values can be used by other clients of the expect test runtime.

val default : t

The default formatting configuration used in expect tests.

module Flexibility : sig ... end