Module Ppx_expect_runtime.For_external

Functions for other libraries to interact with expect tests.

val read_current_test_output_exn : here:Base.Source_code_position.t -> Base.string

If there is an expect test running, perform a consuming read of the current output and return it without any sanitization or backtrace-checking. Note that this is different from the behavior of [%expect.output], which does perform sanitization.

If there is no test running, raise an error that includes here.

val am_running_expect_test : Base.unit -> Base.bool

val default_cr_for_multiple_outputs : output_name:Base.string -> outputs:Base.string Base.list -> Base.string