Module Ppx_expect_runtime.Test_node

type t

module Create : sig ... end

Functions for creating ts corresponding to each of the test nodes that can be parsed out of extension points in a let%expect_test.

module For_mlt : sig ... end

Functions exported for use in toplevel expect tests