Module Test_node.Create

Functions for creating ts corresponding to each of the test nodes that can be parsed out of extension points in a let%expect_test.

Each of these creators accepts the location of the entire AST node associated with the e.g. [%expect] test.

val expect : formatting_flexibility: Ppx_expect_runtime__.Types.Expect_node_formatting.Flexibility.t -> node_loc:Ppx_expect_runtime__.Types.Compact_loc.t -> located_payload: (Ppx_expect_runtime__.Output.Payload.t * Ppx_expect_runtime__.Types.Compact_loc.t) Base.option -> Test_node.t

[%expect _]

val expect_exact : formatting_flexibility: Ppx_expect_runtime__.Types.Expect_node_formatting.Flexibility.t -> node_loc:Ppx_expect_runtime__.Types.Compact_loc.t -> located_payload: (Ppx_expect_runtime__.Output.Payload.t * Ppx_expect_runtime__.Types.Compact_loc.t) Base.option -> Test_node.t

[%expect_exact _]

val expect_unreachable : node_loc:Ppx_expect_runtime__.Types.Compact_loc.t -> Test_node.t

[%expect.unreachable]