Module Ppx_expect_runtime.Write_corrected_file

module Patch_with_file_contents : sig ... end

The callback expected by f, which should convert the input to patches and is allowed to access the contents of the original file while doing so.

val f : use_color:Base.bool -> in_place:Base.bool -> diff_command:Base.string Base.option -> diff_path_prefix:Base.string Base.option -> filename:Base.string -> with_:'a Patch_with_file_contents.t -> 'a -> Ppx_inline_test_lib.Test_result.t

Build a list of diffs to a file using with_, then apply them to the file contents and write a .corrected file.

Exported for use by toplevel tests.