Module Write_corrected_file.Patch_with_file_contents

The callback expected by f, which should convert the input to patches and is allowed to access the contents of the original file while doing so.

type 'a t = original_file_contents:Base.string -> 'a -> (Ppx_expect_runtime__.Types.Compact_loc.t * Base.string) Base.list