Module Ppx_string_runtime.For_string

val empty : string

Empty string.

val of_string : string -> string

Literal string.

val convert : string -> string

Finish a conversion to t.

val concat : string list -> string

Combine multiple values in order.

val pad : string -> len:int -> string

Pad to some minimum length.

val identity : string -> string

Identity function.

Used for ensuring an argument has type t in expanded code, without needing the type t to be exported explicitly for a type annotation. See note above about destructive substitution.