Module type Ppx_string_runtime.S

type t

Result type of interpolation, and of interpolated %

alues}.

type conversion

Result type of %converted#String interpolated values. This will often be either string or t, depending on what is convenient for the configured ppx.

type length

Type of length values for %padding#:8.

val empty : t

Empty string.

val of_string : string -> t

Literal string.

val convert : conversion -> t

Finish a conversion to t.

val concat : t list -> t

Combine multiple values in order.

val pad : t -> len:length -> t

Pad to some minimum length.

val identity : t -> t

Identity function.

Used for ensuring an argument has type t in expanded code, without needing the type t to be exported explicitly for a type annotation. See note above about destructive substitution.